chore: move atom components
