+++ src/app/apply/[recruitmentCode]/_components/apply-selective-box.tsx
@@ -1,6 +1,6 @@
 import { useFormContext } from 'react-hook-form';
-import Container from '../../../../components/shared/container';
-import Typography from '../../../../components/shared/typography';
+import Container from '../../../../components/atom/container.tsx';
+import Typography from '../../../../components/atom/typography.tsx';
 
 import ApplyChoiceBox from './apply-choice-box';
 import { useState } from 'react';
