+++ src/app/recruit/_components/recruit-complete/details/application-detail-dialog.tsx
@@ -1,7 +1,7 @@
 import CrewsDialog from '../../../../../components/molecule/crews-dialog.tsx';
 import useAdminApi from '../../../../../apis/admin-api.ts';
 import { useQuery } from '@tanstack/react-query';
-import Loading from '../../../../../components/shared/loading.tsx';
+import Loading from '../../../../../components/atom/loading.tsx';
 import { throwCustomError } from '../../../../../lib/utils/error.ts';
 import SectionBoxes from '../../../../../components/recruitment-view/section-boxes.tsx';
