+++ src/app/recruit/_components/recruit-make/details/section-boxes.tsx
@@ -7,7 +7,7 @@ import { useToast } from '../../../../../hooks/use-toast.ts';
 import { CREATED_SECTION } from '../../../../../lib/types/default-data.ts';
 import { cn } from '../../../../../lib/utils/utils.ts';
 import useAutosizeTextarea from '../../../../../hooks/use-autosize-textarea.ts';
-import Container from '../../../../../components/shared/container.tsx';
+import Container from '../../../../../components/atom/container.tsx';
 import { z } from 'zod';
 import { CreatedRecruitmentSchema } from '../../../../../lib/types/schemas/recruitment-schema.ts';
