+++ src/app/sign-in/_components/admin-sign-in.tsx
@@ -1,4 +1,4 @@
-import Input from '../../../components/shared/input.tsx';
+import Input from '../../../components/atom/input.tsx';
 import { Button } from '../../../components/shadcn/button.tsx';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { useToast } from '../../../hooks/use-toast.ts';
@@ -8,7 +8,7 @@ import { useNavigate } from 'react-router-dom';
 import useSession from '../../../hooks/use-session.ts';
 import { printCustomError } from '../../../lib/utils/error.ts';
 import { findFirstErrorMessage } from '../../../lib/utils/utils.ts';
-import Loading from '../../../components/shared/loading.tsx';
+import Loading from '../../../components/atom/loading.tsx';
 import {
   isFilledInput,
   isProperClubName,
